#27 Add setting to allow custom output file name
This changeset solves two problems by adding a new setting that allows
users to opt-in to use the "advanced" output file selector.  This new
selector is simply an OpenFileDialog instead of a terrible
FolderBrowserDialog.   The two problems that this new dialog solves are:
(1) A custom file name can be added as requested by issue #27, and (2)
It is much easier to edit the output folder with the OpenFileDialog
because you can type/paste a file path (which helps a bit with issue
#9).